+++ libs/common/src/lib/interfaces/filter.interface.ts
@@ -1,0 +1,5 @@
+export interface Filter {
+  id: string;
+  label?: string;
+  type: 'account' | 'tag';
+}